% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twihttr-package.R
\docType{package}
\name{twihttr}
\alias{twihttr}
\alias{twihttr-package}
\title{A Thin Wrapper of httr to Use Twitter API}
\description{
Simplicity, for great justice.
}
\seealso{
\url{https://dev.twitter.com/rest/public}, \url{https://github.com/hadley/httr/blob/master/demo/oauth1-twitter.r}
}

