% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_remove_profile_banner.R
\name{twtr_account_remove_profile_banner}
\alias{twtr_account_remove_profile_banner}
\title{POST account/remove_profile_banner}
\usage{
twtr_account_remove_profile_banner(...)
}
\description{
POST account/remove_profile_banner
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/account/remove_profile_banner}
}

