% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_update_profile.R
\name{twtr_account_update_profile}
\alias{twtr_account_update_profile}
\title{POST account/update_profile}
\usage{
twtr_account_update_profile(name = NULL, url = NULL, location = NULL,
  description = NULL, profile_link_color = NULL, include_entities = NULL,
  skip_status = NULL, ...)
}
\arguments{
\item{name}{Full name associated with the profile. Maximum of 20 characters.
Example Values: Marcel Molina}

\item{url}{URL associated with the profile. Will be prepended with Åghttp://Åh if not present. Maximum of 100
characters.
Example Values: http://project.ioni.st}

\item{location}{The city or country describing where the user of the account is located. The contents are not
normalized or geocoded in any way. Maximum of 30 characters.
Example Values: San Francisco, CA}

\item{description}{A description of the user owning the account. Maximum of 160 characters.
Example Values: Flipped my wig at age 22 and it never grew back. Also: I work at Twitter.}

\item{profile_link_color}{Sets a hex value that controls the color scheme of links used on the authenticating userÅfs profile
page on twitter.com. This must be a valid hexadecimal value, and may be either three or six
characters (ex: F00 or FF0000).
Example Values: 0000FF
This parameter replaces the deprecated (and separate) update_profile_colors API method.}

\item{include_entities}{The entities node will not be included when set to false.
Example Values: false}

\item{skip_status}{When set to either true, t or 1 statuses will not be included in the returned user objects.}
}
\description{
POST account/update_profile
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/account/update_profile}
}

