% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_update_profile_banner.R
\name{twtr_account_update_profile_banner}
\alias{twtr_account_update_profile_banner}
\title{POST account/update_profile_banner}
\usage{
twtr_account_update_profile_banner(banner, width = NULL, height = NULL,
  offset_left = NULL, offset_top = NULL, ...)
}
\arguments{
\item{banner}{The Base64-encoded or raw image data being uploaded as the userÅfs new profile banner.}

\item{width}{The width of the preferred section of the image being uploaded in pixels. Use with height,
offset_left, and offset_top to select the desired region of the image to use.}

\item{height}{The height of the preferred section of the image being uploaded in pixels. Use with width,
offset_left, and offset_top to select the desired region of the image to use.}

\item{offset_left}{The number of pixels by which to offset the uploaded image from the left. Use with height, width,
and offset_top to select the desired region of the image to use.}

\item{offset_top}{The number of pixels by which to offset the uploaded image from the top. Use with height, width, and
offset_left to select the desired region of the image to use.}
}
\description{
POST account/update_profile_banner
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/account/update_profile_banner}
}

