% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_update_profile_image.R
\name{twtr_account_update_profile_image}
\alias{twtr_account_update_profile_image}
\title{POST account/update_profile_image}
\usage{
twtr_account_update_profile_image(image, include_entities = NULL,
  skip_status = NULL, ...)
}
\arguments{
\item{image}{The avatar image for the profile, base64-encoded. Must be a valid GIF, JPG, or PNG image of less
than 700 kilobytes in size. Images with width larger than 400 pixels will be scaled down. Animated
GIFs will be converted to a static GIF of the first frame, removing the animation.}

\item{include_entities}{The entities node will not be included when set to false.
Example Values: false}

\item{skip_status}{When set to either true, t or 1 statuses will not be included in the returned user objects.}
}
\description{
POST account/update_profile_image
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/account/update_profile_image}
}

