% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_rate_limit_status.R
\name{twtr_application_rate_limit_status}
\alias{twtr_application_rate_limit_status}
\title{GET application/rate_limit_status}
\usage{
twtr_application_rate_limit_status(resources = NULL, ...)
}
\arguments{
\item{resources}{A comma-separated list of resource families you want to know the current rate limit disposition for.
For best performance, only specify the resource families pertinent to your application.
See API Rate Limiting for more information.
Example Values: statuses,friends,trends,help}
}
\description{
GET application/rate_limit_status
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/application/rate_limit_status}
}

