% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks_ids.R
\name{twtr_blocks_ids}
\alias{twtr_blocks_ids}
\title{GET blocks/ids}
\usage{
twtr_blocks_ids(stringify_ids = NULL, cursor = NULL, ...)
}
\arguments{
\item{stringify_ids}{Many programming environments will not consume our ids due to their size. Provide this option to
have ids returned as strings instead. Read more about Twitter IDs.
Example Values: true}

\item{cursor}{Causes the list of IDs to be broken into pages of no more than 5000 IDs at a time. The number of
IDs returned is not guaranteed to be 5000 as suspended users are filtered out after connections are
queried. If no cursor is provided, a value of -1 will be assumed, which is the first Ågpage.Åh The
response from the API will include a previous_cursor and next_cursor to allow paging back and forth.
See Using cursors to navigate collections for more information.
Example Values: 12893764510938}
}
\description{
GET blocks/ids
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/blocks/ids}
}

