% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_messages_destroy.R
\name{twtr_direct_messages_destroy}
\alias{twtr_direct_messages_destroy}
\title{POST direct_messages/destroy}
\usage{
twtr_direct_messages_destroy(id, include_entities = NULL, ...)
}
\arguments{
\item{id}{The ID of the direct message to delete.
Example Values: 1270516771}

\item{include_entities}{The entities node will not be included when set to false.
Example Values: false}
}
\description{
POST direct_messages/destroy
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/direct_messages/destroy}
}

