% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_messages_new.R
\name{twtr_direct_messages_new}
\alias{twtr_direct_messages_new}
\title{POST direct_messages/new}
\usage{
twtr_direct_messages_new(user_id = NULL, screen_name = NULL, text, ...)
}
\arguments{
\item{user_id}{The ID of the user who should receive the direct message. Helpful for disambiguating when a valid
user ID is also a valid screen name.
Example Values: 12345}

\item{screen_name}{The screen name of the user who should receive the direct message. Helpful for disambiguating when a
valid screen name is also a user ID.
Example Values: noradio}

\item{text}{The text of your direct message. Be sure to URL encode as necessary, and keep the message within the
character count limit (available in the help/configuration endpoint)
Example Values: Meet me behind the cafeteria after school}
}
\description{
POST direct_messages/new
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/direct_messages/new}
}

