% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_messages_show.R
\name{twtr_direct_messages_show}
\alias{twtr_direct_messages_show}
\title{GET direct_messages/show}
\usage{
twtr_direct_messages_show(id, ...)
}
\arguments{
\item{id}{The ID of the direct message.
Example Values: 587424932}
}
\description{
GET direct_messages/show
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/direct_messages/show}
}

