% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favorites_create.R
\name{twtr_favorites_create}
\alias{twtr_favorites_create}
\title{POST favorites/create}
\usage{
twtr_favorites_create(id, include_entities = NULL, ...)
}
\arguments{
\item{id}{The numerical ID of the desired status.
Example Values: 123}

\item{include_entities}{The entities node will be omitted when set to false.
Example Values: false}
}
\description{
POST favorites/create
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/favorites/create}
}

