% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favorites_destroy.R
\name{twtr_favorites_destroy}
\alias{twtr_favorites_destroy}
\title{POST favorites/destroy}
\usage{
twtr_favorites_destroy(id, include_entities = NULL, ...)
}
\arguments{
\item{id}{The numerical ID of the desired status.
Example Values: 123}

\item{include_entities}{The entities node will be omitted when set to false.
Example Values: false}
}
\description{
POST favorites/destroy
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/favorites/destroy}
}

