% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followers_list.R
\name{twtr_followers_list}
\alias{twtr_followers_list}
\title{GET followers/list}
\usage{
twtr_followers_list(user_id = NULL, screen_name = NULL, cursor = NULL,
  countoptional = NULL, skip_status = NULL, include_user_entities = NULL,
  ...)
}
\arguments{
\item{user_id}{The ID of the user for whom to return results for.
Example Values: 12345}

\item{screen_name}{The screen name of the user for whom to return results for.
Example Values: twitterdev}

\item{cursor}{Causes the results to be broken into pages. If no cursor is provided, a value of -1 will be assumed,
which is the first Ågpage.Åh The response from the API will include a previous_cursor and next_cursor
to allow paging back and forth. See Using cursors to navigate collections for more information.
Example Values: 12893764510938}

\item{countoptional}{The number of users to return per page, up to a maximum of 200. Defaults to 20.
Example Values: 42}

\item{skip_status}{When set to either true, t or 1, statuses will not be included in the returned user objects. If set
to any other value, statuses will be included.
Example Values: false}

\item{include_user_entities}{The user object entities node will not be included when set to false.
Example Values: false}
}
\description{
GET followers/list
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/followers/list}
}

