% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendships_create.R
\name{twtr_friendships_create}
\alias{twtr_friendships_create}
\title{POST friendships/create}
\usage{
twtr_friendships_create(screen_name = NULL, user_id = NULL, follow = NULL,
  ...)
}
\arguments{
\item{screen_name}{The screen name of the user for whom to befriend.
Example Values: noradio}

\item{user_id}{The ID of the user for whom to befriend.
Example Values: 12345}

\item{follow}{Enable notifications for the target user.
Example Values: true}
}
\description{
POST friendships/create
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/friendships/create}
}

