% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendships_destroy.R
\name{twtr_friendships_destroy}
\alias{twtr_friendships_destroy}
\title{POST friendships/destroy}
\usage{
twtr_friendships_destroy(screen_name = NULL, user_id = NULL, ...)
}
\arguments{
\item{screen_name}{The screen name of the user for whom to unfollow.
Example Values: noradio}

\item{user_id}{The ID of the user for whom to unfollow.
Example Values: 12345}
}
\description{
POST friendships/destroy
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/friendships/destroy}
}

