% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendships_lookup.R
\name{twtr_friendships_lookup}
\alias{twtr_friendships_lookup}
\title{GET friendships/lookup}
\usage{
twtr_friendships_lookup(screen_name = NULL, user_id = NULL, ...)
}
\arguments{
\item{screen_name}{A comma separated list of screen names, up to 100 are allowed in a single request.
Example Values: twitterapi,twitter}

\item{user_id}{A comma separated list of user IDs, up to 100 are allowed in a single request.
Example Values: 783214,6253282}
}
\description{
GET friendships/lookup
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/friendships/lookup}
}

