% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendships_no_retweets_ids.R
\name{twtr_friendships_no_retweets_ids}
\alias{twtr_friendships_no_retweets_ids}
\title{GET friendships/no_retweets/ids}
\usage{
twtr_friendships_no_retweets_ids(stringify_ids = NULL, ...)
}
\arguments{
\item{stringify_ids}{Many programming environments will not consume our ids due to their size. Provide this option to
have ids returned as strings instead. Read more about [node:194]. This parameter is especially
important to use in Javascript environments.
Example Values: true}
}
\description{
GET friendships/no_retweets/ids
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/friendships/no_retweets/ids}
}

