% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendships_outgoing.format.R
\name{twtr_friendships_outgoing.format}
\alias{twtr_friendships_outgoing.format}
\title{GET friendships/outgoing}
\usage{
twtr_friendships_outgoing.format(cursor = NULL, stringify_ids = NULL, ...)
}
\arguments{
\item{cursor}{Causes the list of connections to be broken into pages of no more than 5000 IDs at a time. The
number of IDs returned is not guaranteed to be 5000 as suspended users are filtered out after
connections are queried. If no cursor is provided, a value of -1 will be assumed, which is the
first Ågpage.Åh The response from the API will include a previous_cursor and next_cursor to allow
paging back and forth. See [node:10362, title=ÅhUsing cursors to navigate collectionsÅh] for more
information.
Example Values: 12893764510938}

\item{stringify_ids}{Many programming environments will not consume our Tweet ids due to their size. Provide this option
to have ids returned as strings instead. More about [node:194].
Example Values: true}
}
\description{
GET friendships/outgoing
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/friendships/outgoing}
}

