% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendships_show.R
\name{twtr_friendships_show}
\alias{twtr_friendships_show}
\title{GET friendships/show}
\usage{
twtr_friendships_show(source_id = NULL, source_screen_name = NULL,
  target_id = NULL, target_screen_name = NULL, ...)
}
\arguments{
\item{source_id}{The user_id of the subject user.
Example Values: 3191321}

\item{source_screen_name}{The screen_name of the subject user.
Example Values: raffi}

\item{target_id}{The user_id of the target user.
Example Values: 20}

\item{target_screen_name}{The screen_name of the target user.
Example Values: noradio}
}
\description{
GET friendships/show
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/friendships/show}
}

