% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendships_update.R
\name{twtr_friendships_update}
\alias{twtr_friendships_update}
\title{POST friendships/update}
\usage{
twtr_friendships_update(screen_name = NULL, user_id = NULL, device = NULL,
  retweets = NULL, ...)
}
\arguments{
\item{screen_name}{The screen name of the user for whom to befriend.
Example Values: noradio}

\item{user_id}{The ID of the user for whom to befriend.
Example Values: 12345}

\item{device}{Enable/disable device notifications from the target user.
Example Values: true, false}

\item{retweets}{Enable/disable retweets from the target user.
Example Values: true, false}
}
\description{
POST friendships/update
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/friendships/update}
}

