% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_configuration.R
\name{twtr_help_configuration}
\alias{twtr_help_configuration}
\title{GET help/configuration}
\usage{
twtr_help_configuration(...)
}
\description{
GET help/configuration
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/help/configuration}
}

