% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_privacy.R
\name{twtr_help_privacy}
\alias{twtr_help_privacy}
\title{GET help/privacy}
\usage{
twtr_help_privacy(...)
}
\description{
GET help/privacy
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/help/privacy}
}

