% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_create.R
\name{twtr_lists_create}
\alias{twtr_lists_create}
\title{POST lists/create}
\usage{
twtr_lists_create(name, mode = NULL, description = NULL, ...)
}
\arguments{
\item{name}{The name for the list.
A listÅfs name must start with a letter and can consist only of 25 or fewer letters, numbers, Åg-Åg, or
Åg_Åh characters.}

\item{mode}{Whether your list is public or private. Values can be public or private. If no mode is specified the
list will be public.}

\item{description}{The description to give the list.}
}
\description{
POST lists/create
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/lists/create}
}

