% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_destroy .R
\name{twtr_lists_destroy}
\alias{twtr_lists_destroy}
\title{POST lists/destroy}
\usage{
twtr_lists_destroy(owner_screen_name = NULL, owner_id = NULL, list_id, slug,
  ...)
}
\arguments{
\item{owner_screen_name}{The screen name of the user who owns the list being requested by a slug.}

\item{owner_id}{The user ID of the user who owns the list being requested by a slug.}

\item{list_id}{The numerical id of the list.}

\item{slug}{You can identify a list by its slug instead of its numerical id. If you decide to do so, note that
youÅfll also have to specify the list owner using the owner_id or owner_screen_name parameters.}
}
\description{
POST lists/destroy
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/lists/destroy}
}

