% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_list.R
\name{twtr_lists_list}
\alias{twtr_lists_list}
\title{GET lists/list}
\usage{
twtr_lists_list(user_id = NULL, screen_name = NULL, reverse = NULL, ...)
}
\arguments{
\item{user_id}{The ID of the user for whom to return results for. Helpful for disambiguating when a valid user ID
is also a valid screen name.
Example Values: 12345
Note:: Specifies the ID of the user to get lists from. Helpful for disambiguating when a valid user
ID is also a valid screen name.}

\item{screen_name}{The screen name of the user for whom to return results for. Helpful for disambiguating when a valid
screen name is also a user ID.
Example Values: noradio}

\item{reverse}{Set this to true if you would like owned lists to be returned first. See description above for
information on how this parameter works.
Example Values: true}
}
\description{
GET lists/list
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/lists/list}
}

