% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_members.R
\name{twtr_lists_members}
\alias{twtr_lists_members}
\title{GET lists/members}
\usage{
twtr_lists_members(list_id, slug, owner_screen_name = NULL, owner_id = NULL,
  count = NULL, cursor = NULL, include_entities = NULL,
  skip_status = NULL, ...)
}
\arguments{
\item{list_id}{The numerical id of the list.}

\item{slug}{You can identify a list by its slug instead of its numerical id. If you decide to do so, note that
youÅfll also have to specify the list owner using the owner_id or owner_screen_name parameters.}

\item{owner_screen_name}{The screen name of the user who owns the list being requested by a slug.}

\item{owner_id}{The user ID of the user who owns the list being requested by a slug.}

\item{count}{Specifies the number of results to return per page (see cursor below). The default is 20, with a
maximum of 5,000.}

\item{cursor}{Causes the collection of list members to be broken into ÅgpagesÅh of consistent sizes (specified by
the count parameter). If no cursor is provided, a value of -1 will be assumed, which is the first
Ågpage.Åh The response from the API will include a previous_cursor and next_cursor to allow paging
back and forth. See Using cursors to navigate collections for more information.
Example Values: 12893764510938}

\item{include_entities}{The entities node will be disincluded when set to false.
Example Values: false}

\item{skip_status}{When set to either true, t or 1 statuses will not be included in the returned user objects.}
}
\description{
GET lists/members
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/lists/members}
}

