% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_members_create_all .R
\name{twtr_lists_members_create_all}
\alias{twtr_lists_members_create_all}
\title{POST lists/members/create_all}
\usage{
twtr_lists_members_create_all(list_id, slug, user_id = NULL,
  screen_name = NULL, owner_screen_name = NULL, owner_id = NULL, ...)
}
\arguments{
\item{list_id}{The numerical id of the list.}

\item{slug}{You can identify a list by its slug instead of its numerical id. If you decide to do so, note that
youÅfll also have to specify the list owner using the owner_id or owner_screen_name parameters.}

\item{user_id}{A comma separated list of user IDs, up to 100 are allowed in a single request.}

\item{screen_name}{A comma separated list of screen names, up to 100 are allowed in a single request.}

\item{owner_screen_name}{The screen name of the user who owns the list being requested by a slug.}

\item{owner_id}{The user ID of the user who owns the list being requested by a slug.}
}
\description{
POST lists/members/create_all
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/lists/members/create_all}
}

