% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_members_destroy.R
\name{twtr_lists_members_destroy}
\alias{twtr_lists_members_destroy}
\title{POST lists/members/destroy}
\usage{
twtr_lists_members_destroy(list_id = NULL, slug = NULL, user_id = NULL,
  screen_name = NULL, owner_screen_name = NULL, owner_id = NULL, ...)
}
\arguments{
\item{list_id}{The numerical id of the list.}

\item{slug}{You can identify a list by its slug instead of its numerical id. If you decide to do so, note that
youÅfll also have to specify the list owner using the owner_id or owner_screen_name parameters.}

\item{user_id}{The ID of the user to remove from the list. Helpful for disambiguating when a valid user ID is also
a valid screen name.}

\item{screen_name}{The screen name of the user for whom to remove from the list. Helpful for disambiguating when a
valid screen name is also a user ID.}

\item{owner_screen_name}{The screen name of the user who owns the list being requested by a slug.}

\item{owner_id}{The user ID of the user who owns the list being requested by a slug.}
}
\description{
POST lists/members/destroy
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/lists/members/destroy}
}

