% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_members_show .R
\name{twtr_lists_members_show}
\alias{twtr_lists_members_show}
\title{GET lists/members/show}
\usage{
twtr_lists_members_show(list_id, slug, user_id, screen_name,
  owner_screen_name = NULL, owner_id = NULL, include_entities = NULL,
  skip_status = NULL, ...)
}
\arguments{
\item{list_id}{The numerical id of the list.}

\item{slug}{You can identify a list by its slug instead of its numerical id. If you decide to do so, note that
youÅfll also have to specify the list owner using the owner_id or owner_screen_name parameters.}

\item{user_id}{The ID of the user for whom to return results for. Helpful for disambiguating when a valid user ID
is also a valid screen name.}

\item{screen_name}{The screen name of the user for whom to return results for. Helpful for disambiguating when a valid
screen name is also a user ID.}

\item{owner_screen_name}{The screen name of the user who owns the list being requested by a slug.}

\item{owner_id}{The user ID of the user who owns the list being requested by a slug.}

\item{include_entities}{When set to either true, t or 1, each tweet will include a node called ÅgentitiesÅh. This node offers
a variety of metadata about the tweet in a discreet structure, including: user_mentions, urls, and
hashtags. While entities are opt-in on timelines at present, they will be made a default component
of output in the future. See Tweet Entities for more details.}

\item{skip_status}{When set to either true, t or 1 statuses will not be included in the returned user objects.}
}
\description{
GET lists/members/show
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/lists/members/show}
}

