% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_update.R
\name{twtr_lists_update}
\alias{twtr_lists_update}
\title{POST lists/update}
\usage{
twtr_lists_update(list_id, slug, name = NULL, mode = NULL,
  description = NULL, owner_screen_name = NULL, owner_id = NULL, ...)
}
\arguments{
\item{list_id}{The numerical id of the list.}

\item{slug}{You can identify a list by its slug instead of its numerical id. If you decide to do so, note that
youÅfll also have to specify the list owner using the owner_id or owner_screen_name parameters.}

\item{name}{The name for the list.}

\item{mode}{Whether your list is public or private. Values can be public or private. If no mode is specified the
list will be public.}

\item{description}{The description to give the list.}

\item{owner_screen_name}{The screen name of the user who owns the list being requested by a slug.}

\item{owner_id}{The user ID of the user who owns the list being requested by a slug.}
}
\description{
POST lists/update
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/lists/update}
}

