% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutes_users_create.R
\name{twtr_mutes_users_create}
\alias{twtr_mutes_users_create}
\title{POST mutes/users/create}
\usage{
twtr_mutes_users_create(screen_name = NULL, user_id = NULL, ...)
}
\arguments{
\item{screen_name}{The screen name of the potentially muted user. Helpful for disambiguating when a valid screen name
is also a user ID.
Example Values: whiteleaf}

\item{user_id}{The ID of the potentially muted user. Helpful for disambiguating when a valid user ID is also a
valid screen name.
Example Values: 12345}
}
\description{
POST mutes/users/create
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/mutes/users/create}
}

