% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saved_searches_create.R
\name{twtr_saved_searches_create}
\alias{twtr_saved_searches_create}
\title{POST saved_searches/create}
\usage{
twtr_saved_searches_create(query, ...)
}
\arguments{
\item{query}{The query of the search the user would like to save. The query must be 100 characters or less.}
}
\description{
POST saved_searches/create
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/saved_searches/create}
}

