% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saved_searches_destroy.R
\name{twtr_saved_searches_destroy}
\alias{twtr_saved_searches_destroy}
\title{POST saved_searches/destroy/:id}
\usage{
twtr_saved_searches_destroy(id, ...)
}
\arguments{
\item{id}{The ID of the saved search.
Example Values: 313006}
}
\description{
POST saved_searches/destroy/:id
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/saved_searches/destroy/\%3Aid}
}

