% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saved_searches_list.R
\name{twtr_saved_searches_list}
\alias{twtr_saved_searches_list}
\title{GET saved_searches/list}
\usage{
twtr_saved_searches_list(...)
}
\description{
GET saved_searches/list
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/saved_searches/list}
}

