% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saved_searches_show.R
\name{twtr_saved_searches_show}
\alias{twtr_saved_searches_show}
\title{GET saved_searches/show/:id}
\usage{
twtr_saved_searches_show(id, ...)
}
\arguments{
\item{id}{The ID of the saved search.
Example Values: 313006}
}
\description{
GET saved_searches/show/:id
}

