% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_tweets.R
\name{twtr_search_tweets}
\alias{twtr_search_tweets}
\title{GET search/tweets}
\usage{
twtr_search_tweets(q, ...)
}
\arguments{
\item{q}{A UTF-8, URL-encoded search query of 500 characters maximum, including operators. Queries may additionally be limited by complexity.}
}
\description{
GET search/tweets
}

