% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statuses_destroy.R
\name{twtr_statuses_destroy}
\alias{twtr_statuses_destroy}
\title{POST statuses/destroy/:id}
\usage{
twtr_statuses_destroy(id, trim_user = NULL, ...)
}
\arguments{
\item{id}{The numerical ID of the desired status.
Example Values: 123}

\item{trim_user}{When set to either true, t or 1, each tweet returned in a timeline will include a user object
including only the status authors numerical ID. Omit this parameter to receive the complete user
object.
Example Values: true}
}
\description{
POST statuses/destroy/:id
}

