% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statuses_lookup.R
\name{twtr_statuses_lookup}
\alias{twtr_statuses_lookup}
\title{GET statuses/lookup}
\usage{
twtr_statuses_lookup(idrequired = NULL, include_entities = NULL,
  trim_user = NULL, map = NULL, ...)
}
\arguments{
\item{idrequired}{A comma separated list of tweet IDs, up to 100 are allowed in a single request.
Example Values: 20, 432656548536401920}

\item{include_entities}{The entities node that may appear within embedded statuses will be disincluded when set to false.
Example Values: false}

\item{trim_user}{When set to either true, t or 1, each tweet returned in a timeline will include a user object
including only the status authors numerical ID. Omit this parameter to receive the complete user
object.
Example Values: true}

\item{map}{When using the map parameter, tweets that do not exist or cannot be viewed by the current user will
still have their key represented but with an explicitly null value paired with it
Example Values: true}
}
\description{
GET statuses/lookup
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/statuses/lookup}
}

