% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statuses_retweet.R
\name{twtr_statuses_retweet}
\alias{twtr_statuses_retweet}
\title{POST statuses/retweet/:id}
\usage{
twtr_statuses_retweet(id, trim_user = NULL, ...)
}
\arguments{
\item{id}{The numerical ID of the desired status.
Example Values: 123}

\item{trim_user}{When set to either true, t or 1, each tweet returned in a timeline will include a user object
including only the status authors numerical ID. Omit this parameter to receive the complete user
object.
Example Values: true}
}
\description{
POST statuses/retweet/:id
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/statuses/retweet/\%3Aid}
}

