% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statuses_retweets.R
\name{twtr_statuses_retweets}
\alias{twtr_statuses_retweets}
\title{GET statuses/retweets/:id}
\usage{
twtr_statuses_retweets(id, count = NULL, trim_user = NULL, ...)
}
\arguments{
\item{id}{The numerical ID of the desired status.
Example Values: 123}

\item{count}{Specifies the number of records to retrieve. Must be less than or equal to 100.
Example Values: 5}

\item{trim_user}{When set to either true, t or 1, each tweet returned in a timeline will include a user object
including only the status authors numerical ID. Omit this parameter to receive the complete user
object.
Example Values: true}
}
\description{
GET statuses/retweets/:id
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/statuses/retweets/\%3Aid}
}

