% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statuses_show.R
\name{twtr_statuses_show}
\alias{twtr_statuses_show}
\title{GET statuses/show/:id}
\usage{
twtr_statuses_show(id, trim_user = NULL, include_my_retweet = NULL,
  include_entities = NULL, ...)
}
\arguments{
\item{id}{The numerical ID of the desired Tweet.
Example Values: 123}

\item{trim_user}{When set to either true, t or 1, each tweet returned in a timeline will include a user object
including only the status authors numerical ID. Omit this parameter to receive the complete user
object.
Example Values: true}

\item{include_my_retweet}{When set to either true, t or 1, any Tweets returned that have been retweeted by the authenticating
user will include an additional current_user_retweet node, containing the ID of the source status
for the retweet.
Example Values: true}

\item{include_entities}{The entities node will be disincluded when set to false.
Example Values: false}
}
\description{
GET statuses/show/:id
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/statuses/show/\%3Aid}
}

