% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statuses_unretweet.R
\name{twtr_statuses_unretweet}
\alias{twtr_statuses_unretweet}
\title{POST statuses/unretweet/:id}
\usage{
twtr_statuses_unretweet(id, trim_user = NULL, ...)
}
\arguments{
\item{id}{The numerical ID of the desired retweet status.
Example Values: 123}

\item{trim_user}{When set to either true, t or 1, each tweet returned in a timeline will include a user object
including only the status authors numerical ID. Omit this parameter to receive the complete user
object.
Example Values: true}
}
\description{
POST statuses/unretweet/:id
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/statuses/unretweet/\%3Aid}
}

