% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends_available.R
\name{twtr_trends_available}
\alias{twtr_trends_available}
\title{GET trends/available}
\usage{
twtr_trends_available(...)
}
\description{
GET trends/available
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/trends/available}
}

