% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends_closest.R
\name{twtr_trends_closest}
\alias{twtr_trends_closest}
\title{GET trends/closest}
\usage{
twtr_trends_closest(lat, long, ...)
}
\arguments{
\item{lat}{If provided with a long parameter the available trend locations will be sorted by distance, nearest
to furthest, to the co-ordinate pair. The valid ranges for longitude is -180.0 to +180.0 (West is
negative, East is positive) inclusive.
Example Values: 37.781157}

\item{long}{If provided with a lat parameter the available trend locations will be sorted by distance, nearest
to furthest, to the co-ordinate pair. The valid ranges for longitude is -180.0 to +180.0 (West is
negative, East is positive) inclusive.
Example Values: -122.400612831116}
}
\description{
GET trends/closest
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/trends/closest}
}

