% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends_place.R
\name{twtr_trends_place}
\alias{twtr_trends_place}
\title{GET trends/place}
\usage{
twtr_trends_place(id, exclude = NULL, ...)
}
\arguments{
\item{id}{The Yahoo! Where On Earth ID of the location to return trending information for. Global information
is available by using 1 as the WOEID.
Example Values: 1}

\item{exclude}{Setting this equal to hashtags will remove all hashtags from the trends list.}
}
\description{
GET trends/place
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/trends/place}
}

