% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_lookup.R
\name{twtr_users_lookup}
\alias{twtr_users_lookup}
\title{GET users/lookup}
\usage{
twtr_users_lookup(screen_name = NULL, user_id = NULL,
  include_entities = NULL, ...)
}
\arguments{
\item{screen_name}{A comma separated list of screen names, up to 100 are allowed in a single request. You are strongly
encouraged to use a POST for larger (up to 100 screen names) requests.
Example Values: twitterapi,twitter}

\item{user_id}{A comma separated list of user IDs, up to 100 are allowed in a single request. You are strongly
encouraged to use a POST for larger requests.
Example Values: 783214,6253282}

\item{include_entities}{The entities node that may appear within embedded statuses will be disincluded when set to false.
Example Values: false}
}
\description{
GET users/lookup
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/users/lookup}
}

