% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_profile_banner.R
\name{twtr_users_profile_banner}
\alias{twtr_users_profile_banner}
\title{GET users/profile_banner}
\usage{
twtr_users_profile_banner(user_id = NULL, screen_name = NULL, ...)
}
\arguments{
\item{user_id}{The ID of the user for whom to return results. Helpful for disambiguating when a valid user ID is
also a valid screen name.
Example Values: 12345}

\item{screen_name}{The screen name of the user for whom to return results. Helpful for disambiguating when a valid
screen name is also a user ID.
Example Values: noradio}
}
\description{
GET users/profile_banner
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/users/profile_banner}
}

