% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_report_spam.R
\name{twtr_users_report_spam}
\alias{twtr_users_report_spam}
\title{POST users/report_spam}
\usage{
twtr_users_report_spam(screen_name = NULL, user_id = NULL, ...)
}
\arguments{
\item{screen_name}{The ID or screen_name of the user you want to report as a spammer. Helpful for disambiguating when a
valid screen name is also a user ID.
Example Values: noradio}

\item{user_id}{The ID of the user you want to report as a spammer. Helpful for disambiguating when a valid user ID
is also a valid screen name.
Example Values: 12345}
}
\description{
POST users/report_spam
}
\seealso{
\url{https://dev.twitter.com/rest/reference/post/users/report_spam}
}

