% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_search.R
\name{twtr_users_search}
\alias{twtr_users_search}
\title{GET users/search}
\usage{
twtr_users_search(q, page = NULL, count = NULL, include_entities = NULL,
  ...)
}
\arguments{
\item{q}{The search query to run against people search.
Example Values: Twitter%20API}

\item{page}{Specifies the page of results to retrieve.
Example Values: 3}

\item{count}{The number of potential user results to retrieve per page. This value has a maximum of 20.
Example Values: 5}

\item{include_entities}{The entities node will be disincluded from embedded tweet objects when set to false.
Example Values: false}
}
\description{
GET users/search
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/users/search}
}

