% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_show.R
\name{twtr_users_show}
\alias{twtr_users_show}
\title{GET users/show}
\usage{
twtr_users_show(user_id, screen_name, include_entities = NULL, ...)
}
\arguments{
\item{user_id}{The ID of the user for whom to return results for. Either an id or screen_name is required for this
method.
Example Values: 12345}

\item{screen_name}{The screen name of the user for whom to return results for. Either a id or screen_name is required
for this method.
Example Values: noradio}

\item{include_entities}{The entities node will be disincluded when set to false.
Example Values: false}
}
\description{
GET users/show
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/users/show}
}

