% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_tweets.R
\name{twtr_search_tweets}
\alias{twtr_search_tweets}
\title{GET search/tweets}
\usage{
twtr_search_tweets(q, geocode = NULL, lang = NULL, locale = NULL,
  result_type = NULL, count = NULL, until = NULL, since_id = NULL,
  max_id = NULL, include_entities = NULL, callback = NULL, ...)
}
\arguments{
\item{q}{A UTF-8, URL-encoded search query of 500 characters maximum, including operators. Queries may
additionally be limited by complexity.
Example Values: @noradio}

\item{geocode}{Returns tweets by users located within a given radius of the given latitude/longitude. The location
is preferentially taking from the Geotagging API, but will fall back to their Twitter profile. The
parameter value is specified by Åglatitude,longitude,radiusÅh, where radius units must be specified
as either ÅgmiÅh (miles) or ÅgkmÅh (kilometers). Note that you cannot use the near operator via the API
to geocode arbitrary locations; however you can use this geocode parameter to search near geocodes
directly. A maximum of 1,000 distinct Ågsub-regionsÅh will be considered when using the radius
modifier.
Example Values: 37.781157,-122.398720,1mi}

\item{lang}{Restricts tweets to the given language, given by an ISO 639-1 code. Language detection is best-
effort.
Example Values: eu}

\item{locale}{Specify the language of the query you are sending (only ja is currently effective). This is intended
for language-specific consumers and the default should work in the majority of cases.
Example Values: ja}

\item{result_type}{Optional. Specifies what type of search results you would prefer to receive. The current default
is Ågrecent.Åh Valid values include: <U+00A0><U+00A0>* mixed: Include both popular and real time results in the
response. <U+00A0><U+00A0>* recent: return only the most recent results in the response <U+00A0><U+00A0>* popular: return only
the most popular results in the response.
Example Values: mixed, recent, popular}

\item{count}{The number of tweets to return per page, up to a maximum of 100. Defaults to 15. This was formerly
the ÅgrppÅh parameter in the old Search API.
Example Values: 100}

\item{until}{Returns tweets created before the given date. Date should be formatted as YYYY-MM-DD. Keep in mind
that the search index has a 7-day limit. In other words, no tweets will be found for a date older
than one week.
Example Values: 2015-07-19}

\item{since_id}{Returns results with an ID greater than (that is, more recent than) the specified ID. There are
limits to the number of Tweets which can be accessed through the API. If the limit of Tweets has
occured since the since_id, the since_id will be forced to the oldest ID available.
Example Values: 12345}

\item{max_id}{Returns results with an ID less than (that is, older than) or equal to the specified ID.
Example Values: 54321}

\item{include_entities}{The entities node will be disincluded when set to false.
Example Values: false}

\item{callback}{If supplied, the response will use the JSONP format with a callback of the given name. The
usefulness of this parameter is somewhat diminished by the requirement of authentication for
requests to this endpoint.
Example Values: processTweets}
}
\description{
GET search/tweets
}
\seealso{
\url{https://dev.twitter.com/rest/reference/get/search/tweets}
}

